+++ sbuildr/project/target.py
@@ -1,11 +1,11 @@
-from sbuildr.dependencies.dependency import DependencyLibrary
+from sbuildr.dependencies.dependency import Dependency
 from sbuildr.graph.node import LinkedNode
 from sbuildr.logger import G_LOGGER
 from typing import Dict, List
 
 # A Dict[str, LinkedNode] that maps profile names to the LinkedNodes for a target
 class ProjectTarget(dict):
-    def __init__(self, name, internal: bool=False, dependent_libraries: List[DependencyLibrary]=[], *args, **kwargs):
+    def __init__(self, name, internal: bool=False, dependencies: List[Dependency]=[], *args, **kwargs):
         """
         Represents a single target in a project.
 
@@ -16,8 +16,8 @@ class ProjectTarget(dict):
         self.name = name
         self.internal = False
         self.is_lib = False
-        self.dependent_libraries: List[Dependency] = dependent_libraries
-        G_LOGGER.verbose(f"Created ProjectTarget: {self.name} (internal: {self.internal}), (is_lib: {self.is_lib}), with dependent libraries: {self.dependent_libraries}")
+        self.dependencies: List[Dependency] = dependencies
+        G_LOGGER.verbose(f"Created ProjectTarget: {self.name} (internal: {self.internal}), (is_lib: {self.is_lib}), with dependencies: {self.dependencies}")
 
 
     def __str__(self):