+++ CHANGELOG.md
@@ -29,12 +29,13 @@ Dates are in YYYY-MM-DD format.
 - Adds deferred library propagation. Profiles can now set up libraries just before the build.
 - `Graph` `layers()` now only returns nodes that are actually in the graph.
 - Moves `dependencies` directory up one level.
-- Adds `fetch_dependencies` and `configure_graph` to `Project`
+- Adds `find_dependencies` and `configure_graph` to `Project`
 - Adds `single_dependency` example
 - Adds library name propagation in addition to just `lib_dirs` previously
 - `configure_graph()` is now able to construct partial graphs.
 - `Project` API functions no longer automatically call each other.
 - `SBuildrBuilder` now proapagates `sys.path` to the `PYTHONPATH` environment variable correctly.
+- Adds versioning for `DependencyMetadata` so dependencies in the cache with old versions of Metadata will automatically be rebuilt.
 
 ## v0.4.1 (2019-07-11)
 - Changes generator to favor false positives (longer builds) for `needs_configure()` rather than false negatives (broken builds).